xcpmd: Fix fgets -Werror=unused-result
We have assorted fgets unused-result warnings:
../../git/xcpmd/src/battery.c: In function 'get_num_batteries_present':
../../git/xcpmd/src/battery.c:1003:13: error: ignoring return value of
'fgets', declared with attribute warn_unused_result
[-Werror=unused-result]
             fgets(data, sizeof(data), file);
             ^~~~~~~~~~~~~~~~~

They can be handled by checking for NULL and printing an error.

get_lid_status() changes a little more so we can idicate which file we
failed to read.

The battery.c fgets are inside readdir loops, so we 'continue' to try
reading the next entry.

OXT-1697

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>